--- boilerplates/fe-react/files/src/common/components/Footer.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-
-class Footer extends Component {
-
-  render() {
-    return (
-      <footer>
-        <div className="text-center">
-          Footer Text
-        </div>
-      </footer>
-    )
-  }
-}
-
-export default Footer;
