--- boilerplates/fe-react/files/src/home/actions/actionTypes.js
@@ -1,5 +1,0 @@
-// Request User repos
-export const REQUEST_USER_REPOS_START = 'REQUEST_USER_REPOS_START';
-export const REQUEST_USER_REPOS_SUCCESS = 'REQUEST_USER_REPOS_SUCCESS';
-export const REQUEST_USER_REPOS_FAILED = 'REQUEST_USER_REPOS_FAILED';
-
